Move recipes package into root package
