+++ app/src/main/java/com/example/emmaburton/projectemma/RecipesApplication.kt
@@ -1,4 +1,4 @@
-package com.example.emmaburton.projectemma.presentation.activity
+package com.example.emmaburton.projectemma
 
 import android.app.Application
 import com.facebook.drawee.backends.pipeline.Fresco
