+++ app/src/main/res/layout/activity_recipe_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.activity.RecipeListActivity">
+    tools:context=".recipes.RecipeListActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/search_results_recycler_view"
